refactor experimental namespace; add jax_recipes to autodoc
